+++ extensions/sn_extra_game_options/change_log.md
@@ -8,3 +8,5 @@ Change Log
   - Added option to focus a newly opened map on player
 * 1.2
   - Fixed player-centered map sometimes opening to a generic sector.
+* 1.2
+  - Added failsafe when the map menu alpha fails to find the renderframe.